Day-18/Assignment/lambda/requirements.txt
Pillow==10.4.0
piexif==1.1.3
boto3